requirements.txt
dotenv
streamlit==1.45.0
google-genai ==1.21.1
vertexai
google-cloud
pymongo